+++ src/lib/utils/number.ts
@@ -1,2 +1,2 @@
 export const isPositiveInt = (str: string): boolean =>
-  /^\+?([0-9]\d*)$/.test(str);
+  /^\+?([1-9]\d*)$/.test(str);